refactor: useShakeAnimation 훅 분리
